calendar/requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt[crypto]
django-cors-headers